Free (safe version)

+++ 0x13-more_singly_linked_lists/102-free_listint_safe.c
@@ -10,7 +10,7 @@ size_t free_listint_safe(listint_t **h)
 {
 listint_t *new;
 listint_t *ne;
-size_t x;
+int x;
 x = 0;
 
 new = *h;
